requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
Babel==2.10.3
bcrypt==3.2.2
blinker==1.7.0
Brotli==1.1.0
certifi==2023.11.17
cffi==1.17.1
channels==3.0.4
chardet==5.2.0
click==8.1.6
cloud-init==24.1.3
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
constantly==23.10.4
cryptography==44.0.2
daphne==3.0.2
dbus-python==1.3.2
distro==1.9.0
distro-info==1.7+build1
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.16.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
httplib2==0.20.4
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
legacy-cgi==2.6.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mutagen==1.46.0
netifaces==0.11.0
oauthlib==3.2.2
packaging==24.0
pillow==11.1.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodomex==3.20.0
pycurl==7.45.3
Pygments==2.17.2
PyGObject==3.48.2
PyHamcrest==2.1.0
PyJWT==2.7.0
pyOpenSSL==25.0.0
pyparsing==3.1.1
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu1
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
service-identity==24.2.0
setuptools==78.0.1
six==1.16.0
sqlparse==0.5.3
systemd-python==235
texttable==1.6.7
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
ubuntu-pro-client==8001
unattended-upgrades==0.1
urllib3==2.0.7
wadllib==1.3.6
websocket-client==1.7.0
websockets==10.4
wheel==0.42.0
yt-dlp==2024.4.9
zope.interface==7.2
